refactored settings
