+++ src/containers/Settings/index.ts
@@ -1,0 +1,3 @@
+import Settings from './container';
+
+export default Settings;
