--- src/containers/Settings/index.tsx
@@ -1,3 +1,0 @@
-import SettingsRedux from './Settings-redux';
-
-export default SettingsRedux;
