+++ src/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { StoreContext } from 'redux-react-hook';
 import { AppContainer } from 'react-hot-loader';
 import Root from './containers/Root';
 import { configureStore, history } from './store/configureStore';
@@ -10,10 +9,8 @@ import './styles/app.global.scss';
 const store = configureStore();
 
 ReactDOM.render(
-  <StoreContext.Provider value={store}>
-    <AppContainer>
-      <Root store={store} history={history} />
-    </AppContainer>
-  </StoreContext.Provider>,
+  <AppContainer>
+    <Root store={store} history={history} />
+  </AppContainer>,
   document.getElementById('root')
 );
